% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Bogota_airstations_df}
\alias{Bogota_airstations_df}
\title{Bogota Air Stations Coordinates}
\format{
A data frame with 10 observations and 3 variables:
\describe{
  \item{ESTACION}{Character string representing the name of the station.}
  \item{X}{Numeric value representing the X coordinate of the station.}
  \item{Y}{Numeric value representing the Y coordinate of the station.}
}
}
\source{
Originally taken from the SpatFD package version 0.0.1.
Data provided by the Secretaria de Ambiente de Bogotá:
Monitor network of air quality of Bogota \url{http://rmcab.ambientebogota.gov.co}
}
\usage{
data(Bogota_airstations_df)
}
\description{
This dataset, Bogota_airstations_df, is a data frame containing the coordinates of air quality measurement stations in Bogota, Colombia.
}
\details{
The dataset name has been kept as Bogota_airstations_df to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the ColombiAPI package and assists users in identifying its specific characteristics.
The suffix '_df' indicates that the dataset is a data frame. The original content has not been modified in any way.
}
