% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Bogota_business_Date}
\alias{Bogota_business_Date}
\title{Bogota Business Dates}
\format{
A Date object with 27,173 observations:
\describe{
  \item{Bogota_business_Date}{Date object representing the business dates in Bogota, Colombia, ranging between "1990-01-02" and "2100-12-31".}
}
}
\source{
Originally taken from the quantdates package version 2.0.4.
}
\usage{
data(Bogota_business_Date)
}
\description{
This dataset, Bogota_business_Date, is a Date object containing the business dates for Bogota, Colombia.
}
\details{
The dataset name has been kept as Bogota_business_Date to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the ColombiAPI package and assists users in identifying its specific characteristics.
The suffix '_Date' indicates that the dataset is an object of the Date class. The original content has not been modified in any way.
}
