% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Bogota_holidays_Date}
\alias{Bogota_holidays_Date}
\title{Bogota Holidays Dates}
\format{
A Date object with 1787 observations:
\describe{
  \item{Bogota_holidays_Date}{Date object representing the official holidays in Bogota, Colombia, ranging between "1990-01-01" and "2100-12-08".}
}
}
\source{
Originally taken from the quantdates package version 2.0.4.
Holidays were created using the \code{bizdays} package.
}
\usage{
data(Bogota_holidays_Date)
}
\description{
This dataset, Bogota_holidays_Date, is a Date object containing the official holidays of Bogota, Colombia.
}
\details{
The dataset name has been kept as Bogota_holidays_Date to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the ColombiAPI package and assists users in identifying its specific characteristics.
The suffix '_Date' indicates that the dataset is an object of the Date class. The original content has not been modified in any way.
}
