% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colombiapi-package.R
\docType{package}
\name{ColombiAPI}
\alias{ColombiAPI}
\alias{ColombiAPI-package}
\title{ColombiAPI: Access Colombian Data via APIs and Curated Datasets}
\description{
{
This package provides a comprehensive interface to access diverse public data about Colombia through multiple APIs and curated datasets. The package integrates four different APIs: API-Colombia for Colombian-specific data including geography, culture, tourism, and government information; World Bank API  for economic and demographic indicators; Nager.Date for public holidays; and REST Countries API for general country information.
}
}
\details{
ColombiAPI: Access Colombian Data via APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Colombian Data via APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/colombiapi}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
