% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colombia_literacy_rate.R
\name{get_colombia_literacy_rate}
\alias{get_colombia_literacy_rate}
\title{Get Colombia's Adult Literacy Rate}
\source{
World Bank Open Data API: \url{https://data.worldbank.org/indicator/SE.ADT.LITR.ZS}
}
\usage{
get_colombia_literacy_rate()
}
\value{
A tibble with the following columns:
\itemize{
  \item \code{indicator}: Indicator name
  \item \code{country}: Country name ("Colombia")
  \item \code{year}: Year of the data (integer)
  \item \code{value}: Literacy rate as a percentage
}
}
\description{
Retrieves Colombia's adult literacy rate (% of people ages 15 and above)
for the years 2010 to 2022 using the World Bank Open Data API.
The indicator used is \code{SE.ADT.LITR.ZS}.
}
\details{
This function sends a GET request to the World Bank API.
If the API request fails or returns an error status code,
the function returns \code{NULL} with an informative message.
}
\note{
Requires internet connection.
}
\examples{
if (interactive()) {
  get_colombia_literacy_rate()
}

}
\seealso{
\code{\link[httr]{GET}}, \code{\link[jsonlite]{fromJSON}}, \code{\link[dplyr]{as_tibble}}
}
