% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{municipalities_tbl_df}
\alias{municipalities_tbl_df}
\title{Municipalities of Colombia (Spanish speaking)}
\format{
A tibble with 1,141 observations and 2 variables:
\describe{
  \item{Departamento}{Name of the department (character)}
  \item{Municipio}{Name of the municipality (character)}
}
}
\source{
Data obtained from Kaggle: \url{https://www.kaggle.com/datasets/cavp95/municipios-colombia}
}
\usage{
data(municipalities_tbl_df)
}
\description{
This dataset, municipalities_tbl_df, is a tibble containing the list of municipalities
of Colombia and their corresponding departments, ordered alphabetically.
}
\details{
The dataset name has been kept as 'municipalities_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
ColombiAPI package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble object. The original content has not
been modified in any way.
}
