% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_datasets.R
\name{view_datasets}
\alias{view_datasets}
\title{List of Datasets in the ColombiAPI Package}
\usage{
view_datasets()
}
\value{
A character vector containing the names of the datasets available in the ColombiAPI package.
}
\description{
This function lists all the datasets available in the ColombiAPI package.
It returns a character vector with the names of the datasets included in the package.
}
\examples{
# List the datasets available in the ColombiAPI package
view_datasets()

}
