% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colombiapi-package.R
\docType{package}
\name{ColombiAPI}
\alias{ColombiAPI}
\alias{ColombiAPI-package}
\title{ColombiAPI: Access Colombia's Public Data via API-Colombia}
\description{
{
This package provides access to information about Colombia through the API-Colombia, a public RESTful API that enables users to obtain a variety of public information about Colombia,including general information, geography, and cultural insights.
}
}
\details{
ColombiAPI: Access Colombia's Public Data via API-Colombia

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Colombia's Public Data via API-Colombia.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/colombiapi}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
