% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colombia_attractions.R
\name{get_colombia_attractions}
\alias{get_colombia_attractions}
\title{Get All Touristic Attractions of Colombia}
\usage{
get_colombia_attractions()
}
\value{
A tibble with multiple rows, each representing a touristic attraction in Colombia, or NULL if the API is unavailable.
}
\description{
Retrieves selected information about touristic attractions in Colombia from the API Colombia.
This function fetches the attraction ID, name, description, latitude, longitude, and city ID.
}
\examples{
\donttest{
attractions <- get_colombia_attractions()
print(attractions)
}

}
