% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colombia_invasive_species.R
\name{get_colombia_invasive_species}
\alias{get_colombia_invasive_species}
\title{Get Invasive Species of Colombia}
\usage{
get_colombia_invasive_species()
}
\value{
A tibble with multiple rows and the selected columns, or NULL if the API is unavailable.
}
\description{
Retrieves information about invasive species in Colombia from the API Colombia.
This function fetches only selected fields: id, name, scientificName,
commonNames, impact, manage, and riskLevel.
}
\examples{
\donttest{
invasive_species <- get_colombia_invasive_species()
print(invasive_species)
}

}
