% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colombia_natural_areas.R
\name{get_colombia_natural_areas}
\alias{get_colombia_natural_areas}
\title{Get Natural Areas of Colombia}
\usage{
get_colombia_natural_areas()
}
\value{
A tibble with multiple rows and the selected columns, or NULL if the API is unavailable.
}
\description{
Retrieves information about natural areas in Colombia from the API Colombia.
This function fetches only selected fields: id, areaGroupId, categoryNaturalAreaId,
name, departmentId, daneCode, and landArea.
}
\examples{
\donttest{
natural_areas <- get_colombia_natural_areas()
print(natural_areas)
}

}
