% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colombia_radios.R
\name{get_colombia_radios}
\alias{get_colombia_radios}
\title{Get Radios of Colombia}
\usage{
get_colombia_radios()
}
\value{
A tibble with multiple rows and the selected columns, or NULL if the API is unavailable.
}
\description{
Retrieves information about radio stations in Colombia from the API Colombia.
This function fetches only the following fields: id, name, frequency, and band.
}
\examples{
\donttest{
radios <- get_colombia_radios()
print(radios)
}

}
